requirements.txt
openai
python-dotenv
numpy
chromadb
jupyter
